api_test/requirements.txt
pytest
requests
pyyaml
allure-pytest
pytest-html
pytest-ordering
faker